+++ tests/testthat/test-python-pickle.R
@@ -8,6 +8,6 @@ test_that("Objects can be saved and loaded with pickle", {
   py_save_object(x, "x.pickle")
   on.exit(unlink("x.pickle"), add = TRUE)
   y <- py_load_object("x.pickle")
-  expect_true(x == y)
+  expect_true(py_bool(x == y))
 })